requirements.txt
torch>=2.0.0
gymnasium>=0.28.0
python-chess>=1.9.4
numpy>=1.21.0
